zcash_client_backend: Move the `ShieldedProtocol` and `PoolType` types to `zcash_protocol`
