+++ zcash_client_backend/Cargo.toml
@@ -30,6 +30,7 @@ zcash_encoding.workspace = true
 zcash_keys = { workspace = true, features = ["sapling"] }
 zcash_note_encryption.workspace = true
 zcash_primitives.workspace = true
+zcash_protocol.workspace = true
 zip32.workspace = true
 
 # Dependencies exposed in a public API:
