+++ zcash_keys/Cargo.toml
@@ -22,6 +22,7 @@ rustdoc-args = ["--cfg", "docsrs"]
 zcash_address.workspace = true
 zcash_encoding.workspace = true
 zcash_primitives.workspace = true
+zcash_protocol.workspace = true
 zip32.workspace = true
 
 # Dependencies exposed in a public API:
